rename: admin -> common.admin으로 이동
